backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-docx==1.1.2
openai==1.57.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
aiofiles==24.1.0
sqlalchemy==2.0.0
aiosqlite==0.20.0
